add conditional to unzip gz

+++ src/scripts/unzip-metadata.sh
@@ -1,2 +1,8 @@
-gzip -dk data/gutenberg-metadata.gz \
-    && mv data/gutenberg-metadata data/gutenberg-metadata.json
+FILE=data/gutenberg-metadata.json
+if [ -f "$FILE" ]; then
+    echo "$FILE exists -- no need to unzip"
+else 
+    echo "$FILE does not exist -- unzipping"
+    gzip -dk data/gutenberg-metadata.gz \
+        && mv data/gutenberg-metadata data/gutenberg-metadata.json
+fi
